+++ primitives.cc
@@ -133,7 +133,7 @@ void register_builtin_primitives()
 int get_fragments_from_r(double r, double fn, double fs, double fa)
 {
 	if (fn > 0.0)
-		return fn;
+		return (int)fn;
 	return (int)ceil(fmax(fmin(360.0 / fa, r*M_PI / fs), 5));
 }
